requirements.txt
Flask==1.1.2
Flask-Session
requests
google-cloud-datastore==1.12.0
json2html==1.3.0
google-auth==1.14.3
google-auth-oauthlib
google-auth-httplib2